backend/requirements.txt
Flask==3.0.3 
Flask-SQLAlchemy==3.1.1 
Flask-Cors==4.0.0
Jinja2==3.1.3 
MarkupSafe==2.1.5 
Werkzeug==3.0.2 
blinker==1.8.0 
click==8.1.7 
greenlet==3.0.3 
itsdangerous==2.2.0 
python-dotenv==1.0.1 
sqlalchemy==2.0.29 
psycopg2-binary ==2.9.9
typing-extensions==4.11.0
python-jose==3.3.0

#astroid==2.2.5
#Click==7.0
#ecdsa==0.13.2

#future==0.17.1
#isort==4.3.18
#itsdangerous==1.1.0
#Jinja2==2.10.1
#lazy-object-proxy==1.4.0
#MarkupSafe==1.1.1
#mccabe==0.6.1
#pycryptodome==3.3.1
#pylint==2.3.1
#python-jose-cryptodome==1.3.2
#python-jose
#six==1.12.0
#typed-ast==1.4.2
#Werkzeug==0.15.6
#wrapt==1.11.1
#Flask-Cors==3.0.8
python-dotenv==1.0.1
#used for unit testing
#nose2
#requests